TimeSeries/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
statsmodels==0.14.0
pmdarima==2.0.3
# For ARM64 compatibility, PyTorch should be installed with specific options
# The pip install command in the Dockerfile handles this special case
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2
pyyaml==6.0.1
pytest==7.4.2
plotly==5.16.1
# Additional packages for macOS/M1 compatibility
psutil>=5.9.0 